# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movies - API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/categoriesMovies": {
            "get": {
                "tags": [
                    "CategoriesMovies"
                ],
                "summary": "Get list of categoriesMovies",
                "description": "Returns list of categoriesMovies",
                "operationId": "getCategoriesMoviesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "No categoriesMovie(s) found"
                    }
                }
            },
            "post": {
                "tags": [
                    "CategoriesMovies"
                ],
                "summary": "Store a new categoryMovie",
                "description": "Stores a new categoryMovie and returns the categoryMovie data",
                "operationId": "storeCategoryMovie",
                "requestBody": {
                    "description": "CategoryMovie data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "release_date",
                                    "rating"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A New Beginning"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "It's the story of a new beginning..."
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-09-15"
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "It's the story of a new beginning..."
                                        },
                                        "release_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "rating": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/categoriesMovies/{id}": {
            "get": {
                "tags": [
                    "CategoriesMovies"
                ],
                "summary": "Get categoryMovie by ID",
                "description": "Returns a single categoryMovie",
                "operationId": "getCategoryMovieById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of categoryMovie to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "It's the story of a new beginning..."
                                        },
                                        "release_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "rating": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 8
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CategoryMovie not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "CategoriesMovies"
                ],
                "summary": "Update an existing categoryMovie",
                "description": "Updates and returns a categoryMovie data",
                "operationId": "updateCategoryMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of categoryMovie that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "CategoryMovie data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "release_date",
                                    "rating"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A New Beginning"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "It's the story of a new beginning..."
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-09-15"
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CategoryMovie updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "It's the story of a new beginning..."
                                        },
                                        "release_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "rating": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 8
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "CategoryMovie not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CategoriesMovies"
                ],
                "summary": "Delete a categoryMovie",
                "description": "Deletes a categoryMovie and returns a success message",
                "operationId": "deleteCategoryMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of categoryMovie to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "CategoryMovie deleted successfully"
                    },
                    "404": {
                        "description": "CategoryMovie not found"
                    },
                    "500": {
                        "description": "Error deleting categoryMovie"
                    }
                }
            }
        },
...
    }
}